requirements.txt
cadquery-ocp==7.7.2
selenium==4.23.0
webdriver-manager==4.0.1
